requirements.txt
# PDF processing — no external system dependencies (self-contained)
PyMuPDF>=1.24.0

# Clarifai via OpenAI-compatible endpoint (DeepSeek OCR)
openai>=1.30.0

# Databases
psycopg2-binary>=2.9.9
neo4j>=5.19.0
chromadb>=0.5.0

# Ollama client
ollama>=0.2.0

# Configuration and validation
python-dotenv>=1.0.1
pydantic>=2.7.0

# Utilities
Pillow>=10.3.0
python-dateutil>=2.9.0
